backend/requirements.txt
click==8.0.1
Faker==4.1.1
Flask==1.1.2
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
python-dateutil==2.8.1
six==1.16.0
text-unidecode==1.3
Werkzeug==2.0.1
